% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Markov_trace.R
\name{create_Markov_trace}
\alias{create_Markov_trace}
\title{Create Markov Trace}
\usage{
create_Markov_trace(
  transition_matrix_,
  time_horizon_,
  states_nms_,
  initial_trace_
)
}
\arguments{
\item{transition_matrix_}{Numeric matrix containing the model's transition
matrix.}

\item{time_horizon_}{Numeric scalar defining the number of cycles in the
model.}

\item{states_nms_}{Character vector containing the names of the Markov model
states.}

\item{initial_trace_}{Named numeric vector describing the states' occupancy
in the first cycle of the model.}
}
\value{
A matrix containing the Markov trace.
}
\description{
Create a Markov trace for a State-Transition Model (STM) Markov
providing a transition matrix, time horizon, states' names and the starting
state-occupancy.
}
\examples{
\dontrun{
library("sicksickerPack")
transition_matrix <- matrix(
  data = c(0.845, 0.15, 0, 0.005,
           0.5, 0.3800749, 0.105, 0.01492512,
           0, 0, 0.9511101, 0.04888987,
           0, 0, 0, 1),
  nrow = 4,
  byrow = TRUE,
  dimnames = list(
    c("H", "S1", "S2", "D"),
    c("H", "S1", "S2", "D")
  )
)

Markov_trace <- create_Markov_trace(
  transition_matrix_ = transition_matrix,
  time_horizon_ = 5,
  states_nms_ = c("H", "S1", "S2", "D"),
  initial_trace_ = c("H" = 1, "S1" = 0, "S2" = 0, "D" = 0)
)
}
}
