% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_t_matrix.R
\name{define_t_matrix}
\alias{define_t_matrix}
\title{Define the transition matrix}
\usage{
define_t_matrix(
  p_HD = 0.005,
  p_HS1 = 0.15,
  p_S1H = 0.5,
  p_S1S2 = 0.105,
  hr_S1 = 3,
  hr_S2 = 10
)
}
\arguments{
\item{p_HD}{Numeric, probability to die when healthy.}

\item{p_HS1}{Numeric, probability to become sick when healthy.}

\item{p_S1H}{Numeric, probability to become healthy when sick.}

\item{p_S1S2}{Numeric, probability to become sicker when sick.}

\item{hr_S1}{Numeric, hazard ratio of death in sick vs healthy.}

\item{hr_S2}{Numeric, hazard ratio of death in sicker vs healthy.}
}
\description{
Define the transition matrix of the Sick-Sicker State-Transition
Model (SS-STM).
}
\examples{
\dontrun{
library("sicksickerPack")
define_t_matrix()
}
}
