% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psa.R
\name{run_psa}
\alias{run_psa}
\title{Run Probabilistic Sensitivity Analysis (PSA)}
\usage{
run_psa(
  model_func_ = sicksickerPack::run_sickSicker_model,
  model_func_args_ = list(age_init_ = 25, age_max_ = 55, discount_rate_ = 0.035),
  psa_params_names_ = NULL,
  psa_params_dists_ = NULL,
  psa_params_dists_args_ = NULL,
  n_sim_ = 1000,
  source_url_ = NULL,
  source_path_ = NULL,
  source_credentials_ = NULL
)
}
\arguments{
\item{model_func_}{Function defining the decision analytic model for which
the PSA will be performed.}

\item{model_func_args_}{List of arguments passed to the function named in the
\code{model_func} argument.}

\item{psa_params_names_}{String vector of length one or more defining the
names of the parameters for which PSA configurations will be sampled.}

\item{psa_params_dists_}{String vector of length one or more defining the
names of the distributions of each of the corresponding PSA parameters
named in the vector passed \code{psa_params_names_}.}

\item{psa_params_dists_args_}{List of lists containing named numeric scalar
representing the arguments to be passed to the corresponding PSA
distributions named in the vector passed to the \code{psa_params_dists_} argument.}

\item{n_sim_}{Numeric scalar setting the number of parameter-configurations
to be sampled for PSA purposes.}

\item{source_url_}{String identifying the Uniform Resource Locator (URL)
for the remote (cloud) storage where the function could access a version of
the PSA data. This function expects PSA data in a list which contains three
lists identical to \code{psa_params_names_}, \code{psa_params_dists_} and
\code{psa_params_dists_args_}. The \code{source_url_} should not include a trailing
forward-slash (\code{/}).}

\item{source_path_}{String identifying the path of the data within the URL
passed to the \code{source_url_} argument. If the URL passed to the \code{source_url_}
argument is for an API, then the string passed to this argument is the name
of the API endpoint which returns the data of interest. The \code{source_path_}
should not include a preceding forward-slash (\code{/}).}

\item{source_credentials_}{String identifying the key required to access the
\code{source_url_}. Pass an empty string (\code{""}) if the source does not require a
key, password, or other credentials.}
}
\value{
A data table (dataframe) containing the results from the PSA.
}
\description{
A generic function that performs Probabilistic Sensitivity
Analysis (PSA) given a decision-analytic model.
}
\details{
This function uses the \code{\link{sample_psa_data}} together with
the parameters - passed to the \code{psa_params_names_}, \code{psa_params_dists_}, and
\code{psa_params_dists_args_} arguments to generate PSA parameter configuration.
The sampled PSA sets are then evaluated by the model - passed to the
\code{model_func_} arguments.

Note that this function loops through the PSA parameter sets using the
\code{\link[purrr]{pmap}} function.

Users can choose between passing local data, as in the above example, or
remote PSA data by passing the URL or Application Programming Interface (API)
address, path and any required credentials to the \code{source_url_},
\code{source_path_}, and \code{source_credentials_} arguments. The remote data is then
retrieved by the \code{\link{get_model_params_}} function.
See the examples section for a demo code.
}
\examples{
\dontrun{
library("sicksickerPack")

## Locally stored data:
PSA_results <- run_psa(
  model_func_ = sicksickerPack::run_sickSicker_model,
  model_func_args_ = list(
    age_init_ = 25,
    age_max_  = 55,
    discount_rate_ = 0.035
  ),
  psa_params_names_ = dummy_sickSickerModel_psa_params$
    psa_params_names,
  psa_params_dists_ = dummy_sickSickerModel_psa_params$
    psa_params_dists,
  psa_params_dists_args_ = dummy_sickSickerModel_psa_params$
    psa_params_dists_args,
  n_sim_ = 2
)

## Remotely stored data:
PSA_results <- run_psa(
  model_func_ = sicksickerPack::run_sickSicker_model,
  model_func_args_ = list(
    age_init_ = 25,
    age_max_  = 55,
    discount_rate_ = 0.035
  ),
  n_sim_ = 2,
  source_url_ = "https://sicksickerpack-v7po7ubcwa-nw.a.run.app",
  source_path_ = "psaRunParams",
  source_credentials_ = "R-HTA_2023"
)
}
}
