% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sickSicker_model.R
\name{run_sickSicker_model}
\alias{run_sickSicker_model}
\title{Run the SickSicker State-Transition Model}
\usage{
run_sickSicker_model(
  age_init_ = 25,
  age_max_ = 55,
  discount_rate_ = 0.035,
  p_HD = 0.005,
  p_HS1 = 0.15,
  p_S1H = 0.5,
  p_S1S2 = 0.105,
  hr_S1 = 3,
  hr_S2 = 10,
  c_H = 2000,
  c_S1 = 4000,
  c_S2 = 15000,
  c_Trt = 12000,
  c_D = 0,
  u_H = 1,
  u_S1 = 0.75,
  u_S2 = 0.5,
  u_D = 0,
  u_Trt = 0.95
)
}
\arguments{
\item{age_init_}{Numeric scalar, age at baseline. Default is 25.}

\item{age_max_}{Numeric scalar, maximum age of follow up. Default is 55.}

\item{discount_rate_}{Numeric scalar, discount rate for costs and QALYs.
Default is 0.035.}

\item{p_HD}{Numeric scalar, probability to die when healthy. Default is
0.005.}

\item{p_HS1}{Numeric scalar, probability to become sick when healthy. Default
is 0.15.}

\item{p_S1H}{Numeric scalar, probability to become healthy when sick. Default
is 0.5.}

\item{p_S1S2}{Numeric scalar, probability to become sicker when sick. Default
is 0.105.}

\item{hr_S1}{Numeric scalar, hazard ratio of death in sick vs healthy.
Default is 3.}

\item{hr_S2}{Numeric scalar, hazard ratio of death in sicker vs healthy.
Default is 10.}

\item{c_H}{Numeric scalar, cost of remaining one cycle in the healthy state.
Default is 2000.}

\item{c_S1}{Numeric scalar, cost of remaining one cycle in the sick state.
Default is 4000.}

\item{c_S2}{Numeric scalar, cost of remaining one cycle in the sicker state.
Default is 15000.}

\item{c_Trt}{Numeric scalar, cost of treatment(per cycle). Default is 12000.}

\item{c_D}{Numeric scalar, cost of being in the death state. Default is 0.}

\item{u_H}{Numeric scalar, utility when healthy. Default is 1.}

\item{u_S1}{Numeric scalar, utility when sick. Default is 0.75.}

\item{u_S2}{Numeric scalar, utility when sicker. Default is 0.5.}

\item{u_D}{Numeric scalar, utility when dead. Default is 0.}

\item{u_Trt}{Numeric scalar, utility when being treated. Default is 0.95.}
}
\value{
A matrix containing the costs and QALYs generated by running the
SickSicker model.
}
\description{
Run the SickSicker Markov model.
}
\examples{
\dontrun{
library("sicksickerPack")
sickSicker_model_results <- run_sickSicker_model()
}
}
