% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_psa_data.R
\name{sample_psa_data}
\alias{sample_psa_data}
\title{Sample Probabilistic Sensitivity Analysis (PSA) Configurations}
\usage{
sample_psa_data(
  psa_params_names_,
  psa_params_dists_,
  psa_params_dists_args_,
  n_sim_
)
}
\arguments{
\item{psa_params_names_}{String vector of length one or more defining the
names of the parameters for which PSA configurations will be sampled.}

\item{psa_params_dists_}{String vector of length one or more defining the
names of the distributions of each of the corresponding to the PSA parameters
named in the vector passed \code{psa_params_names_}.}

\item{psa_params_dists_args_}{List of lists containing named numeric scalar
representing the arguments to be passed to the corresponding PSA
distributions named in the vector passed to the \code{psa_params_dists_} argument.}

\item{n_sim_}{Numeric scalar setting the number of parameter-configurations
to be sampled for PSA purposes.}
}
\value{
A data table (dataframe/tibble) containing the parameters'
configurations.
}
\description{
A general purposes function to generate Probabilistic
Sensitivity Analysis (PSA) parameters' configurations. This function allows
sampling from any distribution supported by the \code{base} and \code{truncnorm}
packages - user supplied distributions will be checked against available
functions.
}
\examples{
\dontrun{
library("sicksickerPack")
}
}
