% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_discounting_weights.R
\name{calculate_discounting_weights}
\alias{calculate_discounting_weights}
\title{Calculate discounting weights}
\usage{
calculate_discounting_weights(d_r = 0.035, n_t, first = FALSE)
}
\arguments{
\item{d_r}{Numeric, the rate used to estimate the discounting weights.}

\item{n_t}{Numeric, time horizon, number of cycles for which the discounting
weights are to be calculated.}

\item{first}{Logical, for whether to start discounting from the first cycle.
Default is FALSE, where the discounting weight corresponding to the first
cycle is equal to one.}
}
\value{
A numeric vector containing the discounting weights for
}
\description{
Calculate discounting weights
}
\examples{
\dontrun{
library("sicksickerPack")
}
}
