% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_discounting_weights.R
\name{calculate_discounting_weights}
\alias{calculate_discounting_weights}
\title{Calculate discounting weights}
\usage{
calculate_discounting_weights(
  discount_rate_ = 0.035,
  time_horizon_,
  first_cycle_ = FALSE
)
}
\arguments{
\item{discount_rate_}{Numeric, the rate used to estimate the discounting
weights.}

\item{time_horizon_}{Numeric, time horizon, number of cycles for which the
discounting weights are to be calculated.}

\item{first_cycle_}{Logical, for whether to start discounting from the first
cycle. Default is FALSE, where the discounting weight corresponding to the
first cycle is equal to one.}
}
\value{
A numeric vector containing the discounting weights.
}
\description{
Calculate the discounting weights that correspond to a defined
model time-horizon. Discounting could be employed starting from the initial
cycle.
}
\examples{
\dontrun{
library("sicksickerPack")
calculate_discounting_weights(
  discount_rate_ = 0.035,
  time_horizon_ = 1,
  first_cycle_ = FALSE
)
}
}
