% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_params_.R
\name{get_model_params_}
\alias{get_model_params_}
\title{Grab data hosted remotely}
\usage{
get_model_params_(source_url_, source_path_, source_credentials_)
}
\arguments{
\item{source_url_}{String identifying the Uniform Resource Locator (URL)
for the remote (cloud) storage where the target data is stored. The
\code{source_url_} should not include a trailing forward-slash (\code{/}).}

\item{source_path_}{String identifying the path of the data within the URL
passed to the \code{source_url_} argument. If the URL passed to the \code{source_url_}
argument is for an API, then the string passed to this argument is the name
of the API endpoint which returns the data of interest. The \code{source_path_}
should not include a preceding forward-slash (\code{/}).}

\item{source_credentials_}{String identifying the key required to access the
\code{source_url_}. Pass an empty string (\code{""}) if the source does not require a
key, password, or other credentials.}
}
\value{
An object, likely a data table (dataframe) but depends on the type of
object storing the data in the remote server or returned by the hosting API.
}
\description{
Access and download remotely hosted data. This function allows
access to secured data.
}
\examples{
\dontrun{
library("sicksickerPack")
get_model_params_(
  source_url_ = "https://sicksickerpack-v7po7ubcwa-nw.a.run.app",
  source_path_ = "modelRunParams",
  source_credentials_ = "R-HTA_2023"
)
}
}
