% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psa.R
\name{run_psa}
\alias{run_psa}
\title{Run Probabilistic Sensitivity Analysis (PSA)}
\usage{
run_psa(
  model_func_,
  model_func_args_,
  psa_params_names_,
  psa_params_dists_,
  psa_params_dists_args_,
  n_sim_,
  source_url_ = NULL,
  source_path_ = NULL,
  source_credentials_ = NULL
)
}
\arguments{
\item{model_func_}{String scalar naming the function defining the decision
analytic model.}

\item{model_func_args_}{List of arguments passed to the function named in the
\code{model_func} argument.}

\item{psa_params_names_}{String vector of length one or more defining the
names of the parameters for which PSA configurations will be sampled.}

\item{psa_params_dists_}{String vector of length one or more defining the
names of the distributions of each of the corresponding PSA parameters
named in the vector passed \code{psa_params_names_}.}

\item{psa_params_dists_args_}{List of lists containing named numeric scalar
representing the arguments to be passed to the corresponding PSA
distributions named in the vector passed to the \code{psa_params_dists_}
argument.}

\item{n_sim_}{Numeric scalar setting the number of parameter-configurations
to be sampled for PSA purposes.}

\item{source_url_}{String identifying the Uniform Resource Locator (URL)
for the remote (cloud) storage where the function could access a version of
the PSA data. This function expects PSA data in a list which contains three
lists identical to \code{psa_params_names_}, \code{psa_params_dists_} and
\code{psa_params_dists_args_}.}

\item{source_path_}{String identifying the path of the data within the URL
passed to the \code{source_url_} argument. If the URL passed to the
\code{source_url_} argument is for an API, then the string passed to this
argument is the name of the API endpoint which returns the data of interest.}

\item{source_credentials_}{String identifying the key required to access the
\code{source_url_}.}
}
\value{
A data table (dataframe) containing the results from the PSA.
}
\description{
Run Probabilistic Sensitivity Analysis (PSA)
}
\examples{
\dontrun{
library("sicksickerPack")
PSA_results <- run_psa(
  model_func_ = "run_sickSicker_model",
  model_func_args_ = list(
    age_init_ = 25,
    age_max_  = 55,
    discount_rate_ = 0.035
  ),
  psa_params_names_ = dummy_sickSickerModel_psa_params$
    psa_params_names,
  psa_params_dists_ = dummy_sickSickerModel_psa_params$
    psa_params_dists,
  psa_params_dists_args_ = dummy_sickSickerModel_psa_params$
    psa_params_dists_args,
  n_sim_ = 2
)

PSA_results <- run_psa(
  model_func_ = "run_sickSicker_model",
  model_func_args_ = list(
    age_init_ = 25,
    age_max_  = 55,
    discount_rate_ = 0.035
  ),
  psa_params_names_ = NULL,
  psa_params_dists_ = NULL,
  psa_params_dists_args_ = NULL,
  n_sim_ = 2,
  source_url_ = "http://127.0.0.1:8080",
  source_path_ = "/psaRunParams",
  source_credentials_ = "R-HTA_2023"
)
}
}
