% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sickSicker_model.R
\name{run_sickSicker_model}
\alias{run_sickSicker_model}
\title{Run the SickSicker State-Transition Model}
\usage{
run_sickSicker_model(
  params_ = NULL,
  source_url_ = NULL,
  source_path_ = NULL,
  source_credentials_ = NULL
)
}
\arguments{
\item{params_}{List containing the model parameters. This list should contain
the following objects:
\describe{
\item{age_init_}{Age at baseline, default = \code{25}}
\item{age_max_}{Maximum age of follow up, default = \code{55}}
\item{discount_rate_}{Discount rate for costs and QALYs}
\item{p_HD}{Probability to die when healthy}
\item{p_HS1}{Probability to become sick when healthy}
\item{p_S1H}{Probability to become healthy when sick}
\item{p_S1S2}{Probability to become sicker when sick}
\item{hr_S1}{Hazard ratio of death in sick v healthy}
\item{hr_S2}{Hazard ratio of death in sicker v healthy}
\item{c_H}{Cost of remaining one cycle in the healthy state}
\item{c_S1}{Cost of remaining one cycle in the sick state}
\item{c_S2}{Cost of remaining one cycle in the sicker state}
\item{c_Trt}{Cost of treatment(per cycle)}
\item{c_D}{Cost of being in the death state}
\item{u_H}{Utility when healthy}
\item{u_S1}{Utility when sick}
\item{u_S2}{Utility when sicker}
\item{u_D}{Utility when dead}
\item{u_Trt}{Utility when being treated}
}}

\item{source_url_}{String identifying the Uniform Resource Locator (URL)
for the remote (cloud) storage where the function could access a version of
the data expected in the object passed to the \code{params_} argument. The
\code{source_url_} should not include a trailing forward-slash (\code{/}).}

\item{source_path_}{String identifying the path of the data within the URL
passed to the \code{source_url_} argument. If the URL passed to the \code{source_url_}
argument is for an API, then the string passed to this argument is the name
of the API endpoint which returns the data of interest. The \code{source_path_}
should not include a preceding forward-slash (\code{/}).}

\item{source_credentials_}{String identifying the key required to access the
\code{source_url_}. Pass an empty string (\code{""}) if the source does not require a
key, password, or other credentials.}
}
\value{
A matrix containing the costs and QALYs generated by running the
SickSicker model.
}
\description{
Run the SickSicker Markov model.
}
\details{
Users can choose between passing locally stored parameters to the
\code{params_} argument, as in the first example (below), or remote data by
passing the URL or Application Programming Interface (API) address, path and
any required credentials to the \code{source_url_}, \code{source_path_}, and
\code{source_credentials_} arguments. When required, this function calls the
\code{\link{get_model_params_}} function to retrieve remote data.
See the examples section for a demo code.
}
\examples{
\dontrun{
library("sicksickerPack")

## Locally stored data:
sickSicker_model_results <- run_sickSicker_model(
  params_ = dummy_sickSickerModel_params
)

## Remotely stored data:
run_sickSicker_model(
 source_url_ = "https://sicksickerpack-v7po7ubcwa-nw.a.run.app",
 source_path_ = "modelRunParams",
 source_credentials_ = "R-HTA_2023"
)
}
}
